+++ components/boards/Article/ArticleImage.module.scss
@@ -1,0 +1,11 @@
+@import "@/styles/_color";
+
+.image_container {
+  width: 72px;
+  height: 72px;
+  border-radius: 6px;
+  background-color: $white;
+  border: 1px solid $cool-gray200;
+
+  padding: 12px;
+}
