+++ components/common/LikeCount.module.scss
@@ -1,0 +1,16 @@
+@import "@/styles/_color";
+
+.count_container {
+  display: flex;
+  color: $cool-gray500;
+
+  &.big {
+    gap: 8px;
+    font-size: 16px;
+  }
+
+  &.small {
+    gap: 4px;
+    font-size: 14px;
+  }
+}
